pyproject.toml
[project]
name = "un-website-boilerplate"
version = "0.1.0"
description = "UN Website Boilerplate"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=7.1.0",
    "joblib>=1.5.3",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.10",
    "pymupdf>=1.26.7",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "tqdm>=4.67.1",
]
